Create fscrawler-tika and fscrawler-beans modules
Related to #502.
