+++ beans/src/main/java/fr/pilato/elasticsearch/crawler/fs/meta/doc/Meta.java
