+++ beans/src/main/java/fr/pilato/elasticsearch/crawler/fs/meta/doc/Path.java
