+++ tika/src/main/java/fr/pilato/elasticsearch/crawler/fs/tika/TikaInstance.java
