+++ tika/src/main/java/fr/pilato/elasticsearch/crawler/fs/tika/XmlDocParser.java
