+++ tika/src/test/java/fr/pilato/elasticsearch/crawler/fs/tika/XmlDocParserTest.java
@@ -17,15 +17,23 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.crawler.fs.test.unit.parser;
+package fr.pilato.elasticsearch.crawler.fs.tika;
 
-import fr.pilato.elasticsearch.crawler.fs.test.framework.AbstractFSCrawlerTestCase;
+import org.junit.Test;
 
 import java.io.IOException;
 import java.io.InputStream;
 
-public class DocParserTestCase extends AbstractFSCrawlerTestCase {
-    InputStream getBinaryContent(String filename) throws IOException {
-        return getClass().getResourceAsStream("/documents/" + filename);
+public class XmlDocParserTest extends DocParserTestCase {
+
+    @Test
+    public void testXml() throws IOException {
+        String doc = extractFromFile("issue-163.xml");
+        logger.info("-> {}", doc);
+    }
+
+    private String extractFromFile(String filename) throws IOException {
+        InputStream data = getBinaryContent(filename);
+        return XmlDocParser.generate(data);
     }
 }
